requirements.txt
cryptography
pywin32
mitmproxy
pyinstaller
Pillow
infi.systray